Refactor[]: 변경 서치액티비티를 서치프래그먼트로
