+++ app/src/main/AndroidManifest.xml
@@ -25,8 +25,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".SearchActivity"
-            android:exported="false"/>
     </application>
 
 </manifest>
