+++ app/src/main/res/layout/activity_map.xml
@@ -2,9 +2,10 @@
 <androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_height="match_parent"
-    android:layout_width="match_parent">
+    android:layout_width="match_parent"
+    android:id="@+id/activity_map">
     <FrameLayout
-        android:id="@+id/main"
+        android:id="@+id/activity_map_frameLayout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         tools:context=".MapActivity">
